requirements/prod.txt
flask
flask_sqlalchemy
flask_migrate
redis
geopy
printobject
python-dotenv
confire